Moved util methods and OT from TinyTables to framework to be usable elsewhere in FRESCO
