+++ src/main/java/dk/alexandra/fresco/framework/util/BinaryMatrix.java
@@ -1,4 +1,4 @@
-package dk.alexandra.fresco.suite.tinytables.util;
+package dk.alexandra.fresco.framework.util;
 
 import java.io.Serializable;
 import java.util.BitSet;
