+++ src/main/java/dk/alexandra/fresco/framework/util/ot/OTReceiver.java
@@ -1,8 +1,8 @@
-package dk.alexandra.fresco.suite.tinytables.util.ot;
+package dk.alexandra.fresco.framework.util.ot;
 
 import java.util.List;
 
-import dk.alexandra.fresco.suite.tinytables.util.ot.datatypes.OTSigma;
+import dk.alexandra.fresco.framework.util.ot.datatypes.OTSigma;
 
 public interface OTReceiver {
