+++ src/main/java/dk/alexandra/fresco/framework/util/ot/datatypes/OTSigma.java
@@ -1,4 +1,4 @@
-package dk.alexandra.fresco.suite.tinytables.util.ot.datatypes;
+package dk.alexandra.fresco.framework.util.ot.datatypes;
 
 import java.util.ArrayList;
 import java.util.List;
