+++ src/main/java/dk/alexandra/fresco/framework/util/ot/java/JavaOTFactory.java
@@ -1,11 +1,11 @@
-package dk.alexandra.fresco.suite.tinytables.util.ot.java;
+package dk.alexandra.fresco.framework.util.ot.java;
 
 import java.security.SecureRandom;
 
 import dk.alexandra.fresco.framework.network.Network;
-import dk.alexandra.fresco.suite.tinytables.util.ot.OTFactory;
-import dk.alexandra.fresco.suite.tinytables.util.ot.OTReceiver;
-import dk.alexandra.fresco.suite.tinytables.util.ot.OTSender;
+import dk.alexandra.fresco.framework.util.ot.OTFactory;
+import dk.alexandra.fresco.framework.util.ot.OTReceiver;
+import dk.alexandra.fresco.framework.util.ot.OTSender;
 
 public class JavaOTFactory implements OTFactory {
