+++ src/main/java/dk/alexandra/fresco/framework/util/ot/java/JavaOTSender.java
@@ -1,4 +1,4 @@
-package dk.alexandra.fresco.suite.tinytables.util.ot.java;
+package dk.alexandra.fresco.framework.util.ot.java;
 
 import java.io.IOException;
 import java.security.SecureRandom;
@@ -6,9 +6,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import dk.alexandra.fresco.framework.network.Network;
-import dk.alexandra.fresco.suite.tinytables.util.Encoding;
-import dk.alexandra.fresco.suite.tinytables.util.ot.OTSender;
-import dk.alexandra.fresco.suite.tinytables.util.ot.datatypes.OTInput;
+import dk.alexandra.fresco.framework.util.ot.Encoding;
+import dk.alexandra.fresco.framework.util.ot.OTSender;
+import dk.alexandra.fresco.framework.util.ot.datatypes.OTInput;
 import edu.biu.scapi.exceptions.FactoriesException;
 import edu.biu.scapi.exceptions.SecurityLevelException;
 import edu.biu.scapi.interactiveMidProtocols.ot.otBatch.OTBatchOnByteArraySInput;
