+++ src/main/java/dk/alexandra/fresco/suite/tinytables/online/datatypes/TinyTablesSBool.java
@@ -26,8 +26,8 @@
  *******************************************************************************/
 package dk.alexandra.fresco.suite.tinytables.online.datatypes;
 
+import dk.alexandra.fresco.framework.util.ot.Encoding;
 import dk.alexandra.fresco.framework.value.SBool;
-import dk.alexandra.fresco.suite.tinytables.util.Encoding;
 
 /**
  * This class represents a masked boolean value in the online phase of the
