+++ src/main/java/dk/alexandra/fresco/suite/tinytables/prepro/datatypes/TinyTablesPreproSBool.java
@@ -26,8 +26,8 @@
  *******************************************************************************/
 package dk.alexandra.fresco.suite.tinytables.prepro.datatypes;
 
+import dk.alexandra.fresco.framework.util.ot.Encoding;
 import dk.alexandra.fresco.framework.value.SBool;
-import dk.alexandra.fresco.suite.tinytables.util.Encoding;
 
 /**
  * This class represents a masked boolean value in the preprocessing phase of
